Use right joystick for precision firing on PS3
Reworked the input_axis and axis_is_active logic for PS3; the right
joystick controls the input_axis, and axis_is_active is set when it
is not centered.  This allows the use the right joystick for aiming
independently of the left joystick used for movement.  Also, activate
the fire command when axis_is_active.